requirements.txt
djangorestframework == 3.15.1
django-cors-headers == 4.3.1 
djangorestframework-simplejwt == 5.3.1


